chore: split code in two modules for readability
